+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     * Fix: Reset pending Orders and Complete processing Orders will no longer send emails.
     * Fix: Video played via prettyPhoto now loads properly.
     * Fix: shipping_dropdown.php and payment_methods.php now can be replaced in theme files.
+    * Fix: Product total price after ajax update in cart now is calculated properly.
 * 1.14 - 2014-12-11:
     * New: Used Coupon column on Orders page.
     * New: Email variables.